Add changes to upload port information from fea stitch to cp fea. Read the port information from the system in fea stitch and forward it to fea.
Also, add persistence to the fea-stitch slot number to uid mapping by
storing it in a file that the cp fea will manage.
Similarly, also store the port num to interface names in a file for persistence.
This file will be maanged by fea-stitch.

Lastly, fea stitch will also store its uid in a file for persistence.